Refactor travelled-countries-counter
